backend/requirements.txt
# Core Framework
fastapi==0.117.1
uvicorn[standard]==0.36.0
pydantic==2.11.9
pydantic-settings==2.10.1
aiohttp==3.12.15

# Database
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
psycopg2-binary==2.9.10
asyncpg==0.30.0
aiosqlite==0.21.0
redis==6.4.0

# Vector Database
qdrant-client==1.15.1

# AI Agents and Frameworks
crewai==0.193.2
pydantic-ai==1.0.8

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
authlib==1.3.2
itsdangerous==2.2.0
email-validator==2.3.0
hvac==2.3.0

# HTTP Client & Resilience
httpx==0.28.1
tenacity==9.1.2

# Observability
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-instrumentation-httpx==0.58b0
opentelemetry-instrumentation-sqlalchemy==0.58b0
opentelemetry-instrumentation-redis==0.58b0

# Streaming
sse-starlette==3.0.2

# Background Tasks
# celery[redis]==5.6.0b1  # Removed due to Redis version conflict

# Content Processing
textstat==0.7.10
markdown==3.9

# Development & Testing
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
black==25.9.0
isort==6.0.1
flake8==7.3.0
mypy==1.18.2
ruff==0.8.3

# Utilities
python-dotenv==1.1.1
structlog==25.4.0
click==8.3.0
PyYAML==6.0.2
